requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130

flask-cors==6.0.1
Flask==3.1.2
accelerate==1.11.0
awscli==1.42.61
black==25.9.0
boto3==1.40.61
datasets==2.21.0
jinja2==3.1.6
git+https://github.com/felipemaiapolo/tinyBenchmarks
huggingface-hub==0.36.0
hydra-core==1.3.2
ipykernel==7.1.0
ipython==8.37.0
isort==7.0.0
jupyterlab==4.4.10
lm_eval==0.4.9.1
notebook==7.4.7
omegaconf==2.3.0
psutil==7.1.3
pytest==8.4.2
python-terraform==0.10.1
ragas==0.3.2
s3fs==2024.6.1
sentence-transformers==5.1.2
tensorboard==2.20.0
torch==2.9.1+cu130 
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
trl==0.15.2
wandb==0.22.3
watchdog==6.0.0
faker==37.12.0

